store: Split the test-store into modules
The main motiviation for this is so that we can build everything in the
test-store conditioned on `debug_assertions`. Trying to do this
item-by-item is becoming too difficult.
